services/ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
PyJWT==2.8.0
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
python-multipart==0.0.9
httpx==0.27.0
